# packages/api-client/openapi.json
{
  "components": {
    "responses": {
      "DEFAULT_ERROR": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Default error response"
      },
      "UNPROCESSABLE_ENTITY": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Unprocessable Entity"
      }
    },
    "schemas": {
...
      "OperationalDataEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "additionalProperties": {},
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "TransactionDetailEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "TransactionHistoryEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionHistoryData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "TransactionUploadEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionUploadData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MoMo-FDVS API",
    "version": "v1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/admin/transactions": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "maxLength": 30,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalDataEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "operations-v1"
        ]
      }
    },
    "/api/v1/admin/transactions/{transaction_id}": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalDataEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "operations-v1"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "transaction_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ]
    },
...
    "/api/v1/transactions": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "DRAFT",
                "UPLOADED",
                "OCR_PENDING",
                "OCR_REVIEW",
                "READY",
                "ANALYSIS_QUEUED",
                "ANALYSING",
                "COMPLETED",
                "PARTIAL",
                "FAILED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "verification",
            "required": false,
            "schema": {
              "enum": [
                "VERIFIED",
                "UNVERIFIED",
                "MISMATCH",
                "NOT_ATTEMPTED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "band",
            "required": false,
            "schema": {
              "enum": [
                "low_risk",
                "medium_risk",
                "high_risk",
                "inconclusive"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionHistoryEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "List only the caller's transactions using persisted latest-run summaries.",
        "tags": [
          "transactions-v1"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "client_captured_at": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "client_metadata": {
                    "description": "A size-limited JSON object of safe scalar values.",
                    "type": "string"
                  },
                  "receipt": {
                    "format": "binary",
                    "type": "string"
                  },
                  "source": {
                    "enum": [
                      "CAMERA",
                      "GALLERY"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "receipt",
                  "source"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionUploadEnvelope"
                }
              }
            },
            "description": "A completed idempotent upload was replayed."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionUploadEnvelope"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Invalid image metadata or idempotency key."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The USER role is required."
          },
          "409": {
            "description": "Idempotency conflict."
          },
          "413": {
            "description": "Encoded bytes or decoded pixels exceed a limit."
          },
          "415": {
            "description": "Unsupported, corrupt, disguised, or unsafe image."
          },
          "429": {
            "description": "Per-user upload rate limit exceeded."
          },
          "503": {
            "description": "Private storage is unavailable."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Create a transaction from a strictly validated private receipt.",
        "tags": [
          "transactions-v1"
        ]
      }
    },
    "/api/v1/transactions/{transaction_id}": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetailEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The USER role is required."
          },
          "404": {
            "description": "Transaction not found."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Return a bounded owner-safe transaction and immutable run history.",
        "tags": [
          "transactions-v1"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "transaction_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ]
    },
...
    "/api/v1/transactions/{transaction_id}/ocr": {
      "parameters": [
        {
          "in": "path",
          "name": "transaction_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ],
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRReviewEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Missing or invalid idempotency key."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The USER role is required."
          },
          "404": {
            "description": "Transaction not found."
          },
          "409": {
            "description": "Invalid state, idempotency conflict or hash mismatch."
          },
          "429": {
            "description": "Per-user OCR rate limit exceeded."
          },
          "503": {
            "description": "Private storage or OCR processing is unavailable."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Run a bounded, reconstructable OCR pipeline for an owned receipt.",
        "tags": [
          "ocr-v1"
        ]
      }
    },
...
    "/api/v1/transactions/{transaction_id}/receipt": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "default": "thumbnail",
              "enum": [
                "thumbnail",
                "original",
                "ela",
                "noise-map"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Private receipt image bytes."
          },
          "404": {
            "description": "Receipt not found or not visible to the caller."
          },
          "503": {
            "description": "Private storage is unavailable."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Stream an authorised original, thumbnail or staff-only diagnostic.",
        "tags": [
          "transactions-v1"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "transaction_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ]
    },
...
  },
...
}